+++ lib/iex/api.rb
@@ -19,6 +19,7 @@ end
 
 require_relative 'api/chart'
 require_relative 'api/company'
+require_relative 'api/dividends'
 require_relative 'api/logo'
 require_relative 'api/key_stats'
 require_relative 'api/news'